% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_get_events.R
\name{List Available Events}
\alias{List Available Events}
\alias{mb_get_events}
\title{Get List of Available Events}
\usage{
mb_get_events(session_data, start_date = Sys.Date(), end_date = Sys.Date() +
  2, sport_ids = c(15), market_states = c("open", "suspended"))
}
\arguments{
\item{session_data}{A session object returned from a successful mb_login attempt. It contains details about your user preferences and security details.}

\item{start_date}{A string (or date/POSIXct) value with format YYYY-mm-dd or YYYY-mm-dd HH:MM:SS format. Defaults to events starting today .}

\item{end_date}{A string (or date/POSIXct) value with format YYYY-mm-dd or YYYY-mm-dd HH:MM:SS format. Defaults to events starting before two days time.}

\item{sport_ids}{A vector of integer sport_ids that indicated sports for which event details are required. e.g. c(15,1) gives Soccer and Pro Football (NFL)}

\item{market_states}{A vector of string containing the market states to return. Defaults to 'open' or 'suspended' market types.}
}
\value{
If successful, a dataframe with first 500 events and associated information. Only 500 events are permitted at one time. Pagination is possible but not implemented in this version.
}
\description{
List the Events Available on Matchbook.com
}
\examples{
\dontrun{mb_get_events(sport_ids=15)}
}
\seealso{
\code{\link{mb_get_sports}}
}

