% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_get_markets.R
\name{List Available Markets for an Event}
\alias{List Available Markets for an Event}
\alias{mb_get_markets}
\title{Get List of Available Markets for a given Event}
\usage{
mb_get_markets(event_id, market_states = c("open", "suspended"),
  market_types = c("multirunner", "binary"),
  grading_types = c("asian-handicap", "high-score-wins", "low-score-wins",
  "point-spread", "point-total", "single-winner-wins"),
  include_runners = FALSE, include_prices = FALSE, language = "en")
}
\arguments{
\item{event_id}{The event_id integer for which a list of associated markets is required.}

\item{market_states}{A vector of string containing the market states to return. Defaults to 'open' or 'suspended' market types.}

\item{market_types}{A vector of strings containing the required market types. Valid market types are either 'multirunner' or 'binary'. Both are returned by default.}

\item{grading_types}{A vector of strings containing the required grading types. Valid grading types are 'asian-handicap','high-score-wins','low-score-wins','point-spread','point-total','single-winner-wins'. All are returned by default.}

\item{include_runners}{A boolean parameter indicating if the runner names and id should be returned or not. Defaults to FALSE.}

\item{include_prices}{A boolean parameter indicating if the runner prices should be returned or not. Defaults to FALSE.}

\item{language}{A string with the language parameter e.g. 'en'}

\item{sport_ids}{A vector of integer sport_ids that indicated sports for which event details are required. e.g. c(15,1) gives Soccer and Pro Football (NFL)}
}
\value{
If successful, a dataframe with first 500 markets and associated information. Only 500 markets are permitted at one time. Pagination is possible but not implemented here.
}
\description{
List the Markets Available on Matchbook.com for a given Event
}
\examples{
\dontrun{mb_get_markets(event_id=309912)}
}
\seealso{
\code{\link{mb_get_sports},\link{mb_get_events}}
}

