% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_get_sports.R
\name{List Available Sports}
\alias{List Available Sports}
\alias{mb_get_sports}
\title{Get List of Available Sports}
\usage{
mb_get_sports(session_data, status = "active")
}
\arguments{
\item{session_data}{A session object returned from a successful mb_login attempt. It contains details about your user preferences and security details.}

\item{status}{A string with value one of 'active', 'pending'}
}
\value{
If successful, the list of returned sports data.
}
\description{
List the Sports Available on Matchbook.com
}
\examples{
\dontrun{mb_get_sports()}
}
\seealso{
\code{\link{mb_get_events}}
}

