% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_logout.R
\name{Logout and Session termination}
\alias{Logout and Session termination}
\alias{mb_logout}
\title{Logout of the Matchbook.com API session}
\usage{
mb_logout(username, user_id, session_token)
}
\arguments{
\item{username}{Your Matchbook.com username. If you are not already registered got to: https://www.matchbook.com/}

\item{user_id}{Your Matchbook.com user_id obtained from the mb_login call.}

\item{session_token}{Your Matchbook.com session_token obtained from the mb_login call.}
}
\value{
The response insteger status_code.
}
\description{
End the Matchbook.com session.
}
\examples{
\dontrun{mb_logout("my_username","my_user_id","my_private_session_token")}
}
\seealso{
\code{\link{mb_login}}
}

