\name{matchbook-package}
\alias{matchbook-package}
\alias{matchbook}
\docType{package}
\title{
\packageTitle{matchbook}
}
\description{
\packageDescription{matchbook}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{matchbook}
\packageIndices{matchbook}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{matchbook}

Maintainer: \packageMaintainer{matchbook}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
