% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_bet_place.R
\name{mb_bet_place}
\alias{mb_bet_place}
\title{Perform a Bet Placement Action}
\usage{
mb_bet_place(session_data, runner_id, side, stake, odds)
}
\arguments{
\item{session_data}{A session object returned from a successful mb_login attempt. It contains details about your user preferences and security details.}

\item{runner_id}{The id of the runner that you want to place a bet on.}

\item{side}{The side that you want place a bet. This has to be one of either 'back' or 'lay'.}

\item{stake}{The amount that want to stake for this bet. The currency used is the currency you specified when you set up your Matchbook.com account. A real number format is required.}

\item{odds}{The odds you want to place a bet at. The odds type is based on the information from your session_data, which is the default setting saved for your account.}
}
\value{
The status and details of your bet placement are returned.
}
\description{
This function provide the core bet placement functionality. Its possible to place a single bet by entering in a single value for each of the runner_id, side, odds and stake parameters. It also possible to place multiple bets at once by passing a vector for each of the runner_id, side, odds and stake parameters. In this instance, its imperative that the order in each of the vectors is preserved i.e. that the runner_id, side, odds and stake for the first bet are the first elements of each parameter vector, and that the runner_id, side, odds and stake values for the second bet are the second elements of each parameter vector and so on. There is a cap of 20 on the number of bets that can be placed at one time.
}
\examples{
\dontrun{my_session <- mb_login("my_user_name","verysafepassword");
mb_bet_place(session_data=my_session,odds=2.5,stake=5,runner_id=12345)}
}
\seealso{
\code{\link{mb_get_bets}}
}

