% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_get_bets.R
\name{mb_get_bets}
\alias{mb_get_bets}
\title{Get List of Current Bets on Matchbook}
\usage{
mb_get_bets(session_data, event_ids = NULL, market_ids = NULL,
  runner_ids = NULL, sides = NULL, status = NULL, interval = 0)
}
\arguments{
\item{session_data}{A session object returned from a successful mb_login attempt. It contains details about your user preferences and security details.}

\item{event_ids}{A vector of event_ids for which a list of current bets is required. This is an optional parameter and the default is to return bets from all events unless market_ids or runner_ids are specified.}

\item{market_ids}{A vector of market_ids for which a list of current bets is required. This is an optional parameter and the default is to return bets from all markets unless event_ids or runner_ids are specified.}

\item{runner_ids}{A vector of runner_ids for which a list of current bets is required. This is an optional parameter and the default is to return bets from all runners unless event_ids or market_ids are specified.}

\item{sides}{A filter to allow selection of either 'back' or 'lay' bets. The default is to return both types.}

\item{status}{The bet status from one of the possible options ('matched','unmatched','cancelled','expired','opened','paused'). By default matched and unmatched bets are returned.}

\item{interval}{Time filter (in seconds) to allow selection of bets that were created or updated in the period between the currnet time and the current time minus the specified number of seconds.}
}
\value{
If successful, a dataframe with first 500 bets and associated information. Only 500 bets are permitted at one time. Pagination is possible but not implemented in this version.
}
\description{
List the first 500 bets that have been made on Matchbook events that have not yet settled.
}
\examples{
\dontrun{my_session <- mb_login("my_user_name","my_password");
mb_get_bets(session_data=my_session)}
}
\seealso{
\code{\link{mb_get_sports},\link{mb_get_events},\link{mb_get_markets}}
}

