% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_get_currencies.R
\name{mb_get_currencies}
\alias{mb_get_currencies}
\title{Get List of Available Currencies}
\usage{
mb_get_currencies(session_data)
}
\arguments{
\item{session_data}{A session object returned from a successful \code{\link{mb_login}} attempt. It contains security details and your account preferences.}
}
\value{
If successful, a data frame with the following fields:
 \describe{
  \item{currency-id}{currency name}
  \item{short-name}{short version of currency name}
  \item{long-name}{long version of currency name}
  }
}
\description{
List the Currencies currently available on Matchbook.com.
}
\examples{
\dontrun{my_session <- mb_login("my_user_name","versafepassword"); 
mb_get_currencies(session_data=my_session)}

}
\seealso{
\code{\link{mb_login}}
}

