% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_get_settled.R
\name{mb_get_settled}
\alias{mb_get_settled}
\title{Get Details of Settled Bets}
\usage{
mb_get_settled(session_data, sport_id = NULL, period = NULL,
  start_date = Sys.Date() - 90, end_date = Sys.Date())
}
\arguments{
\item{session_data}{A session object returned from a successful mb_login attempt. It contains details about your user preferences and security details.}

\item{period}{A value with one of 'today', '1-day', '2-day', 'yesterday', 'week', 'month', '1-month', '3-month'}

\item{start_date}{A string (or date/POSIXct) value with format YYYY-mm-dd or YYYY-mm-dd HH:MM:SS format. Defaults to 90 days ago.}

\item{end_date}{A string (or date/POSIXct) value with format YYYY-mm-dd or YYYY-mm-dd HH:MM:SS format. Defaults to today.}

\item{sport_ids}{A vector of integer sport_ids that indicated sports for which event details are required. e.g. c(15,1) gives Soccer and Pro Football (NFL)}
}
\value{
If successful, a dataframe with first 500 settled bets. Only 500 bets are permitted at one time. Pagination is possible but not implemented in this version.
}
\description{
Get Details of all Settled Bets
}
\examples{
\dontrun{my_session <- mb_login("my_user_name","verysafepassword");
mb_get_settled(session_data=my_session,event_id=309912)}
}
\seealso{
\code{\link{mb_get_bets},\link{mb_get_events}}
}

