% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_login.R
\name{mb_login}
\alias{mb_login}
\title{Login and Authenticate with the Matchbook.com API}
\usage{
mb_login(username, password)
}
\arguments{
\item{username}{Your Matchbook.com username. If you are not already registered got to: https://www.matchbook.com/}

\item{password}{The password provided to you by email by the Matchbook.com team.}
}
\value{
A list with status_code, user_id and session_token values for your session.
}
\description{
Validate your session with the Matchbook API.
}
\examples{
\dontrun{mb_login("my_username","verysafepassword")}
}
\seealso{
\code{\link{mb_logout}}
}

