% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_logout.R
\name{mb_logout}
\alias{mb_logout}
\title{Terminate a Matchbook.com API session}
\usage{
mb_logout(session_data)
}
\arguments{
\item{session_data}{This is a required parameter containing security and preference information. It must take the exact format of the object returned by the mb_login request.}
}
\value{
The response integer status_code.
}
\description{
End the Matchbook.com session.
}
\examples{
\dontrun{my_session <- mb_login("my_user_name","verysafepassword"); 
mb_logout(my_session)}

}
\seealso{
\code{\link{mb_login}}
}

