% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_bet_cancel.R
\name{Perform an Action to Cancel a Single or Multiple Exchange Bets}
\alias{Perform an Action to Cancel a Single or Multiple Exchange Bets}
\alias{mb_bet_cancel}
\title{Perform a Bet Cancel Action}
\usage{
mb_bet_cancel(session_data, offer_id = NULL, event_id = NULL,
  market_id = NULL, runner_id = NULL, cancel_all = FALSE)
}
\arguments{
\item{session_data}{A session object returned from a successful mb_login attempt. It contains details about your user preferences and security details.}

\item{offer_id}{The offer_id or vector of offer_ids that you want to cancel.}

\item{event_id}{The event_id or vector of event_ids that you want to cancel.}

\item{market_id}{The market_id or vector of market_ids that you want to cancel.}

\item{runner_id}{The runner_id or vector of runner_ids that you want to cancel.}

\item{cancel_all}{Boolean variable. Parameter to allow cancellation of all bets on all events/markets/runners. Default is FALSE.}
}
\value{
The status and details of your bet cancellations are returned.
}
\description{
This function provides bet cancellation functionality. Its possible to cancel a single bet by entering in a single value for the offer_id parameter. It also possible to cancel multiple bets at once by passing a vector of the offer_id parameter. Its also possible to cancel all bets from a given market, event or runner or vector of market, event or runner.
}
\examples{
\dontrun{my_session <- mb_login("my_user_name","verysafepassword");
mb_bet_cancel(session_data=my_session,odds=2.5,stake=5,runner_id=12345)}
}
\seealso{
\code{\link{mb_get_bets},\link{mb_bet_place},\link{mb_bet_updatel}}
}

