% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mb_logout.R
\name{Logout and Session termination}
\alias{Logout and Session termination}
\alias{mb_logout}
\title{Logout of the Matchbook.com API session}
\usage{
mb_logout(session_data)
}
\arguments{
\item{session_data}{This is a required paramter contianin security and preference information. Its an object that must take the exact format of an mb_login request.}
}
\value{
The response insteger status_code.
}
\description{
End the Matchbook.com session.
}
\examples{
\dontrun{my_session <- mb_login("my_user_name","verysafepassword");
mb_logout(my_session)}
}
\seealso{
\code{\link{mb_login}}
}

